carlasimu/requirements.txt
absl-py==1.4.0
alabaster==0.7.13
albumentations==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.13.2
atomicwrites==1.4.1
attrs==22.2.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.12.0
bleach==5.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
cookiecutter==2.1.1
cryptography==39.0.0
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
docker==6.0.1
docopt==0.6.2
docstring-to-markdown==0.11
docutils==0.19
entrypoints==0.4
fastjsonschema==2.16.2
flake8==5.0.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
imageio==2.25.1
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isort==5.11.4
jaraco.classes==3.2.3
jedi==0.18.2
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==23.13.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
more-itertools==9.0.0
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
numpydoc==1.5.0
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pafy==0.5.5
pandas==1.3.5
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.2.3
pyflakes==3.0.1
pygame==2.1.3
Pygments==2.14.0
pylint==2.15.10
pylint-venv==2.3.0
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.0
python-slugify==7.0.0
pytoolconfig==1.2.4
pytz==2022.7.1
PyWavelets==1.3.0
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
QDarkStyle==3.0.3
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rope==1.6.0
rsa==4.9
Rtree==1.0.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.1
spyder-kernels==2.4.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
text-unidecode==1.3
textdistance==4.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
three-merge==0.1.1
tifffile==2021.11.2
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==0.11.0
torchsummary==1.5.1
torchvision==0.14.1+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typed-ast==1.5.4
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
watchdog==2.2.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
whatthepatch==1.0.3
widgetsnbextension==4.0.5
wrapt==1.14.1
yapf==0.32.0
yarg==0.1.9
youtube-dl==2020.12.2
zipp==3.11.0
